fix(ui): allow users to disable the camera after initial load (#165)
This commit fixes a bug that caused the camera to always revert to the
default, preventing users from disabling it. This fix ensures that the default
camera is only selected on the first initial load, and users can disable it
afterwards. It also applies a quick patch to warn people if no camera or microphone
is selected.