add ComfyUI-FasterLivePortrait, models (#142)
* add ComfyUI-FasterLivePortrait, models

* whitespace

* chore: update FasterLivePortrait node branch

Updates the FasterLivePortrait repository branch to the `main` branch
now that the latest changes have been merged in upstream.

* fix: fix broken FasterLivePortrait branch name

Ensures that the `master` branch is used since the `main` branch does
not excists.

* include onnx for liveportrait trt, update build deps

* add liveportrait trt build deps

* update model paths for fasterliveportrait

* add build step to entrypoint.sh, fix dl paths

* update node branch in config

* fix onnx model urls

* pin tensorrt bindings in comfyui-base

* fix joyvasa model urls

* handle huggingface redirects in downloads

* fix: ensure faster LivePortrait engines are built

A small bug in `entrypoint.sh` caused the faster LivePortrait engine build
step to be skipped. The script was checking the wrong directory to determine
if the engines already existed. This commit corrects the logic to ensure the
engines are built properly.

* update faster live portrait for test

* update branch for FasterLivePortrait

---------

Co-authored-by: Rick Staa <rick.staa@outlook.com>
Co-authored-by: John | Elite Encoder <john@eliteencoder.net>